requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai>=1.50.0,<3.0.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.2
httpx==0.27.2
streamlit==1.39.0
pandas==2.2.3
altair==5.4.1
python-multipart==0.0.12
